Replace int with long int and long with long long

+++ examples/Application/Builder/MainApplication.cpp
@@ -12,6 +12,6 @@ public:
 	}
 };
 
-int main(int argc, char **argv) {
+long int main(long int argc, char **argv) {
 	return Application(MainApplication::main, argc, argv);
 }
