+++ java/lang/Bytes/Bytes.hpp
@@ -62,7 +62,7 @@ namespace Java {
             /**
              * The number of bytes used to represent a byte value in two's complement binary form.
              */
-            static constexpr int BYTES = 1;
+            static constexpr long int BYTES = 1;
 
         public:
             /**
@@ -131,7 +131,7 @@ namespace Java {
              * a value less than 0 if this Byte is numerically less than other;
              * and a value greater than 0 if this Byte is numerically greater than other
              */
-            int compareTo(const Bytes &other) const override;
+            long int compareTo(const Bytes &other) const override;
 
             /**
              * Decodes a String into a Byte. Accepts decimal, hexadecimal, and octal numbers
@@ -170,7 +170,7 @@ namespace Java {
              *
              * @return long
              */
-            int hashCode() const;
+            long int hashCode() const;
 
             /**
              * Returns a hash code for this Byte
@@ -185,14 +185,14 @@ namespace Java {
              *
              * @return int
              */
-            int intValue() const override;
+            long int intValue() const override;
 
             /**
              * Returns the value of this Byte as an long.
              *
              * @return long
              */
-            long longValue() const override;
+            long long longValue() const override;
 
             /**
              * Parse the parameter string as a byte with radix =10
@@ -211,7 +211,7 @@ namespace Java {
              * @throw NumberFormatException If the string does not contain a parsable byte.
              * @return the byte value represented by the string argument in the specified radix
              */
-            static byte parseBytes(String stringToParse, int radix);
+            static byte parseBytes(String stringToParse, long int radix);
 
             /**
              * Returns the value of this Byte as an short.
@@ -236,7 +236,7 @@ namespace Java {
             static String toString(byte byteValue);
 
             /**
-             * Converts the argument to an int by an unsigned conversion.
+             * Converts the argument to an long int by an unsigned conversion.
              *
              * @param byteValue
              * @return int
@@ -244,12 +244,12 @@ namespace Java {
             static long int toUnsignedInt(byte byteValue);
 
             /**
-             * Converts the argument to an int by an unsigned conversion.
+             * Converts the argument to an long int by an unsigned conversion.
              *
              * @param byteValue
              * @return long
              */
-            static long toUnsignedLong(byte byteValue);
+            static long long toUnsignedLong(byte byteValue);
 
             /**
              * Returns a Byte instance representing the specified byte value.
@@ -274,7 +274,7 @@ namespace Java {
              * @param radix
              * @return Byte
              */
-            static Bytes valueOf(String stringValue, int radix);
+            static Bytes valueOf(String stringValue, long int radix);
 
         public:
             /**
@@ -439,7 +439,7 @@ namespace Java {
 
             Array<Bytes> cacheInit() {
                 Array<Bytes> cacheArray;
-                int index;
+                long int index;
                 for (index = 0; index < 256; index++) {
                     cacheArray.push(Bytes(static_cast<byte>(index)));
                 }
@@ -454,7 +454,7 @@ namespace Java {
                 return instance;
             }
 
-            Bytes getByteAtIndex(int index) {
+            Bytes getByteAtIndex(long int index) {
                 return cache[index];
             }
         };
